+++ libs/docs/core/user-menu/examples/user-menu-mobile-example.component.ts
@@ -56,6 +56,8 @@ export class UserMenuMobileExampleComponent {
 
     expanded = true;
     isOpen = false;
+    menuItems = Array.from({ length: 20 }, (_, i) => `Menu Item ${i + 1}`);
+    subMenuItems = Array.from({ length: 20 }, (_, i) => `Sub-menu Item ${i + 1}`);
 
     constructor(private _messageToastService: MessageToastService) {}