+++ libs/i18n/src/lib/translations/translations_ro.ts
@@ -1,4 +1,3 @@
-/* eslint-disable */
 // Do not modify, it's automatically created. Modify translations_ro.properties instead
 export default {
     coreCalendar: {
@@ -26,7 +25,8 @@ export default {
     },
     coreMultiComboBox: {
         multiComboBoxAriaLabel: 'Casetă combinată cu valori multiple',
-        selectAllLabel: 'Selectare toate ({selectedItems} din {totalItems})'
+        selectAllLabel: 'Selectare toate ({selectedItems} din {totalItems})',
+        listGroupHeader: 'Group Header'
     },
     coreCarousel: {
         pageIndicatorLabel: 'Item {itemNum} of {totalNum} displayed',
@@ -612,5 +612,9 @@ export default {
         groupAriaDescriptionCollapsed: 'restrâns',
         triggerMoreLabel: 'Mai mult',
         triggerLessLabel: 'Mai puțin'
+    },
+    segmentedButton: {
+        groupRoleDescription: 'Segmented Button Group',
+        buttonRoleDescription: 'Segmented Button'
     }
 };